internal: Hoise and fix normalizr examples
