+++ examples/normalizr-github/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "normalizr-github-example",
+  "version": "0.0.0",
+  "description": "And example of using Normalizr with github",
+  "main": "index.js",
+  "author": "Paul Armstrong",
+  "license": "MIT",
+  "private": true,
+  "scripts": {
+    "start": "babel-node ./ --root-mode upward --extensions '.ts,.tsx,.js'"
+  },
+  "dependencies": {
+    "@rest-hooks/normalizr": "^7.4.0"
+  }
+}
