+++ examples/normalizr-redux/README.md
@@ -1,0 +1,15 @@
+# Redux
+
+This is a simple example of using [@rest-hooks/normalizr](../packages/normalizr) with Redux and Redux-Thunk. The command-line utility allows you to pull some data from public GitHub repos and browse/display the normalized data as saved to the Redux state tree.
+
+![redux example in use](/packages/normalizr/examples/redux/usage.gif)
+
+## Running
+
+From this directory, run the following and follow the on-screen options:
+
+```sh
+# from the root directory:
+yarn # or npm install
+npm run start
+```
