+++ examples/normalizr-redux/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "normalizr-redux-example",
+  "version": "0.0.0",
+  "description": "And example of using Normalizr with Redux",
+  "main": "index.js",
+  "author": "Paul Armstrong",
+  "license": "MIT",
+  "private": true,
+  "scripts": {
+    "start": "babel-node ./ --root-mode upward --extensions '.ts,.tsx,.js'"
+  },
+  "devDependencies": {
+    "@octokit/rest": "^18.6.0",
+    "@rest-hooks/normalizr": "^7.4.0",
+    "inquirer": "^8.1.1",
+    "redux": "^4.1.0",
+    "redux-thunk": "^2.3.0"
+  }
+}
