+++ examples/normalizr-redux/src/api/index.js
@@ -1,0 +1,7 @@
+import { Octokit } from '@octokit/rest';
+
+export default new Octokit({
+  headers: {
+    'user-agent': 'Normalizr Redux Example',
+  },
+}).rest;
