+++ examples/normalizr-redux/src/redux/index.js
@@ -1,0 +1,11 @@
+import thunk from 'redux-thunk';
+import { applyMiddleware, createStore } from 'redux';
+
+import * as schema from '../api/schema';
+import api from '../api';
+import reducer from './reducer';
+
+export default createStore(
+  reducer,
+  applyMiddleware(thunk.withExtraArgument({ api, schema })),
+);
