+++ examples/normalizr-relationships/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "normalizr-relationships",
+  "version": "0.0.0",
+  "description": "And example of using Normalizr with relationships",
+  "main": "index.js",
+  "author": "Paul Armstrong",
+  "license": "MIT",
+  "private": true,
+  "scripts": {
+    "start": "babel-node ./ --root-mode upward --extensions '.ts,.tsx,.js'"
+  },
+  "devDependencies": {
+    "mockdate": "^3.0.5"
+  },
+  "dependencies": {
+    "@rest-hooks/normalizr": "^7.4.0"
+  }
+}
