--- packages/normalizr/examples/redux/.babelrc
@@ -1,3 +1,0 @@
-{
-  "presets": ["es2015", "stage-1"]
-}
