--- packages/normalizr/examples/redux/README.md
@@ -1,17 +1,0 @@
-# Redux
-
-This is a simple example of using Normalizr with Redux and Redux-Thunk. The command-line utility allows you to pull some data from public GitHub repos and browse/display the normalized data as saved to the Redux state tree.
-
-![redux example in use](/examples/redux/usage.gif)
-
-## Running
-
-From this directory, run the following and follow the on-screen options:
-
-```sh
-# from the root directory:
-yarn # or npm install
-# from this directory:
-yarn # or npm install
-npm run start
-```
