--- packages/normalizr/examples/redux/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "normalizr-redux-example",
-  "version": "0.0.0",
-  "description": "And example of using Normalizr with Redux",
-  "main": "index.js",
-  "author": "Paul Armstrong",
-  "license": "MIT",
-  "private": true,
-  "scripts": {
-    "start": "babel-node ./"
-  },
-  "dependencies": {
-    "babel-cli": "^6.18.0",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-stage-1": "^6.16.0",
-    "github": "^14.0.0",
-    "inquirer": "^8.1.1",
-    "redux": "^4.1.0",
-    "redux-thunk": "^2.1.0"
-  }
-}
