--- packages/normalizr/examples/redux/src/api/index.js
@@ -1,7 +1,0 @@
-import GitHubApi from 'github';
-
-export default new GitHubApi({
-  headers: {
-    'user-agent': 'Normalizr Redux Example'
-  }
-});
