--- packages/normalizr/examples/redux/src/redux/index.js
@@ -1,7 +1,0 @@
-import * as schema from '../api/schema';
-import api from '../api';
-import reducer from './reducer';
-import thunk from 'redux-thunk';
-import { applyMiddleware, createStore } from 'redux';
-
-export default createStore(reducer, applyMiddleware(thunk.withExtraArgument({ api, schema })));
